// src/react/modules/interface.ts
export interface headerMenuu {
    label: string
}
export interface feature {
    logo: any
    text: string
    price: string
    title: string
    btn: string
}
export interface upComing {
    title: string
    text: string
    btn: string
}
export interface recent {
    logo: any
    date: string
    title: string
    text: string
}